// src/users/enums/state.enum.ts
export enum State {
  Alabama = 'al',
  Alaska = 'ak',
  Arizona = 'az',
  Arkansas = 'ar',
  California = 'ca',
  Colorado = 'co',
  Connecticut = 'ct',
  Delaware = 'de',
  Florida = 'fl',
  Georgia = 'ga',
  Hawaii = 'hi',
  Idaho = 'id',
  Illinois = 'il',
  Indiana = 'in',
  Iowa = 'ia',
  Kansas = 'ks',
  Kentucky = 'ky',
  Louisiana = 'la',
  Maine = 'me',
  Maryland = 'md',
  Massachusetts = 'ma',
  Michigan = 'mi',
  Minnesota = 'mn',
  Mississippi = 'ms',
  Missouri = 'mo',
  Montana = 'mt',
  Nebraska = 'ne',
  Nevada = 'nv',
  NewHampshire = 'nh',
  NewJersey = 'nj',
  NewMexico = 'nm',
  NewYork = 'ny',
  NorthCarolina = 'nc',
  NorthDakota = 'nd',
  Ohio = 'oh',
  Oklahoma = 'ok',
  Oregon = 'or',
  Pennsylvania = 'pa',
  RhodeIsland = 'ri',
  SouthCarolina = 'sc',
  SouthDakota = 'sd',
  Tennessee = 'tn',
  Texas = 'tx',
  Utah = 'ut',
  Vermont = 'vt',
  Virginia = 'va',
  Washington = 'wa',
  WestVirginia = 'wv',
  Wisconsin = 'wi',
  Wyoming = 'wy',
}